requirements.txt
ultralytics 
opencv-python 
pillow
numpy